test(e2e): split vue cases to separate folder
