+++ e2e/cases/vue/jsx-basic/index.test.ts
@@ -1,0 +1,17 @@
+import { expect } from '@playwright/test';
+import { build, getHrefByEntryName } from '@scripts/shared';
+import { rspackOnlyTest } from '@scripts/helper';
+
+rspackOnlyTest('should build basic Vue jsx correctly', async ({ page }) => {
+  const rsbuild = await build({
+    cwd: __dirname,
+    runServer: true,
+  });
+
+  await page.goto(getHrefByEntryName('index', rsbuild.port));
+
+  const button1 = page.locator('#button1');
+  await expect(button1).toHaveText('A: 0');
+
+  await rsbuild.close();
+});
