+++ e2e/tsconfig.json
@@ -16,5 +16,6 @@
     "playwright.config.ts",
     "cases/**/*.test.ts",
     "uni-builder/**/*.test.ts"
-  ]
+  ],
+  "exclude": ["**/node_modules", "**/.*/"]
 }
